modificar tipo de variavel "Int" para Float

+++ app/src/main/java/com/example/geek_central/component/NoteComponent.kt
@@ -31,7 +31,7 @@ class NoteComponent(val view : View) {
         setMaxSeekBar(valueInital)
     }
 
-    fun setValueComponentTextNote(value : Int = 0){
+    fun setValueComponentTextNote(value : Float = 0F){
         txtNote.text = value.toString()
     }
